app/requirements.txt
appdirs==1.4.3
defusedxml==0.5.0
Django==1.10.5
django-allauth==0.31.0
django-cors-headers==2.0.2
django-crispy-forms==1.6.1
django-extensions==1.7.7
django-pandas==0.4.1
django-pyodbc-azure==1.10.4.0
djangorestframework==3.6.2
lazy-property==0.0.1
numpy==1.12.0+mkl
oauthlib==2.0.2
packaging==16.8
pandas==0.19.2
pickleshare==0.7.4
pyodbc==4.0.15
pyparsing==2.1.10
python-dateutil==2.6.0
python3-openid==3.1.0
pytz==2016.10
requests==2.13.0
requests-oauthlib==0.8.0
scikit-learn==0.18
scipy==0.19.0rc1
simplejson==3.10.0
six==1.10.0
